+++ sdk/communication/azure-communication-sms/src/samples/java/com/azure/communication/sms/samples/quickstart/ReadmeSamples.java
@@ -12,7 +12,7 @@ import com.azure.core.credential.AzureKeyCredential;
 import com.azure.core.credential.TokenCredential;
 import com.azure.core.http.HttpClient;
 import com.azure.core.http.netty.NettyAsyncHttpClientBuilder;
-import com.azure.core.http.rest.PagedIterable;
+import com.azure.core.http.rest.Response;
 import com.azure.core.util.Context;
 import com.azure.identity.DefaultAzureCredentialBuilder;
 
@@ -117,12 +117,12 @@ public class ReadmeSamples {
         options.setDeliveryReportEnabled(true);
         options.setTag("Marketing");
 
-        Iterable<SmsSendResult> sendResults = smsClient.send(
+        Iterable<SmsSendResult> sendResults = smsClient.sendWithResponse(
             "<from-phone-number>",
             Arrays.asList("<to-phone-number1>", "<to-phone-number2>"),
             "Weekly Promotion",
             options /* Optional */,
-            Context.NONE);
+            Context.NONE).getValue();
 
         for (SmsSendResult result : sendResults) {
             System.out.println("Message Id: " + result.getMessageId());
@@ -155,13 +155,14 @@ public class ReadmeSamples {
             options.setDeliveryReportEnabled(true);
             options.setTag("Marketing");
 
-            PagedIterable<SmsSendResult> smsSendResults = smsClient.send(
+            Response<Iterable<SmsSendResult>> sendResults = smsClient.sendWithResponse(
                 "<from-phone-number>",
                 Arrays.asList("<to-phone-number1>", "<to-phone-number2>"),
                 "Weekly Promotion",
                 options /* Optional */,
                 Context.NONE);
 
+            Iterable<SmsSendResult> smsSendResults = sendResults.getValue();
             for (SmsSendResult result : smsSendResults) {
                 if (result.isSuccessful()) {
                     System.out.println("Successfully sent this message: " + result.getMessageId() + " to " + result.getTo());