+++ sdk/resourcemanager/azure-resourcemanager-servicebus/CHANGELOG.md
@@ -1,6 +1,6 @@
 # Release History
 
-## 2.3.0-beta.1 (Unreleased)
+## 2.3.0 (2021-03-30)
 
 - Updated `api-version` to `2017-04-01`
 - Parameter of `regenerateKey` method changed from `Policykey` to `KeyType`